feat: reorganize exception imports to user package for improved clarity
