+++ backend/src/main/java/com/cinerama/backend/service/impl/AuthServiceImpl.java
@@ -6,8 +6,8 @@ import com.cinerama.backend.dto.RegisterRequest;
 import com.cinerama.backend.dto.VerificationRequest;
 import com.cinerama.backend.entity.Role;
 import com.cinerama.backend.entity.User;
-import com.cinerama.backend.exception.PasswordsNotMatchException;
-import com.cinerama.backend.exception.UserNotFoundException;
+import com.cinerama.backend.exception.user.PasswordsNotMatchException;
+import com.cinerama.backend.exception.user.UserNotFoundException;
 import com.cinerama.backend.repository.RoleRepository;
 import com.cinerama.backend.repository.UserRepository;
 import com.cinerama.backend.service.AuthService;
