+++ backend/src/main/java/com/cinerama/backend/service/impl/MailServiceImpl.java
@@ -4,7 +4,7 @@ import com.cinerama.backend.service.MailService;
 import com.cinerama.backend.service.mail.MailContentBuilder;
 import com.cinerama.backend.entity.User;
 import com.cinerama.backend.repository.UserRepository;
-import com.cinerama.backend.exception.UserNotFoundException;
+import com.cinerama.backend.exception.user.UserNotFoundException;
 import jakarta.mail.MessagingException;
 import jakarta.mail.internet.MimeMessage;
 import org.springframework.beans.factory.annotation.Autowired;
